pyproject.toml
[project]
name = "stock-pro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "discord-py>=2.6.4",
    "django>=6.0.1",
    "django-apscheduler>=0.7.0",
    "django-environ>=0.12.0",
    "django-filter>=25.2",
    "finance-datareader>=0.9.102",
    "google-genai>=1.60.0",
    "gunicorn>=24.1.1",
    "html5lib>=1.1",
    "lxml>=6.0.2",
    "matplotlib>=3.10.8",
    "pandas<3.0.0",
    "pandas-ta>=0.4.71b0",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
    "yfinance>=1.1.0",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "isort>=7.0.0",
]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip = ["migrations"]
